LPS-75775 Move to internal
