+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/java/com/liferay/dynamic/data/lists/form/web/internal/constants/DDLFormPortletKeys.java
@@ -12,14 +12,20 @@
  * details.
  */
 
-package com.liferay.dynamic.data.lists.form.web.constants;
+package com.liferay.dynamic.data.lists.form.web.internal.constants;
 
 /**
- * @author Lino Alves
+ * @author Marcellus Tavares
  */
-public class DDLFormConstants {
+public class DDLFormPortletKeys {
 
-	public static final String SERVICE_NAME =
-		"com.liferay.dynamic.data.lists.form.web";
+	public static final String DYNAMIC_DATA_LISTS_FORM =
+		"com_liferay_dynamic_data_lists_form_web_portlet_DDLFormPortlet";
+
+	public static final String DYNAMIC_DATA_LISTS_FORM_ADMIN =
+		"com_liferay_dynamic_data_lists_form_web_portlet_DDLFormAdminPortlet";
+
+	public static final String DYNAMIC_DATA_LISTS_FORM_BROWSER =
+		"com_liferay_dynamic_data_lists_form_web_portlet_DDLFormBrowserPortlet";
 
 }
